pyproject.toml
[tool.poetry]
name = "crowd-test-client"
version = "0.1.0"
description = ""
authors = ["Викторов Андрей Германович <Viktorov.A.G@sberbank.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
aioboto3 = "^9.6.0"
click = "^8.1.3"
tqdm = "^4.64.0"
aio-pika = "^7.2.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
crowd = 'crowd_cli:main'
